+++ src/interfaces/auth.ts
@@ -1,0 +1,6 @@
+export interface AuthState {
+  username: string | null;
+  userId: string | null;
+  accessToken: string | null;
+  refreshToken: string | null;
+}
