+++ src/interfaces/textbook.ts
@@ -1,0 +1,4 @@
+export interface TextbookState {
+  group: number;
+  page: number;
+}
