+++ src/store/auth/authSlice.ts
@@ -1,12 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import type { AuthState } from '../../interfaces/auth';
 import type { RootState } from '..';
-
-export interface AuthState {
-  username: string | null;
-  userId: string | null;
-  accessToken: string | null;
-  refreshToken: string | null;
-}
 
 const initialState: AuthState = {
   username: null,
